Finished basic machine functionality
The gacha machine can now accept an item and provide another item in exchange. The machine also properly drops it's item when broken in survival and does not in creative. The machine has a 2 block tall hitbox and model. The machine faces the player regardless of the direction from which the player placed the machine.